Add per candidate stateful caching and lookup of possible winner status.
Under the assumption of non-reversion of manipulation for a candidate,
we can save calls to the sat solver by short-circuiting when it has
been seen that a candidate can be made to win with fewer manipulators,
or cannot be made to win with more.  Cache this info in a stateful
cache kept with unsafePeformIO and MVars.