Big overhaul of API for NPLib.
Renamed the NPrograms and NProgramComputations to Instance and
InstanceBuilder respectively.  You now code up instances in the
InstanceBuilder monad, then call buildInstance to get an Instance
record.  Accessing the fields of the Instance record will lazily solve
the formula underlying the instance, and give just the requested
information (formula satisfiability, deterministic solutions for the
nvars under the models, comments/stats from the sat solver regarding
the solve attempt.  Solving an instance for just one model consists
of accessing head of a lazily-computed list of solutions.

Introduced a functional dependency in the Interpret class, eliminating
the need to put explicit type signatures on the returned values and
the need to pass `show' into `ntrace'.

Updated all clients in this repo to use the new API.

Temporarily disabled ZChaff, Minisat and RSat solvers.